refactor: modularize api routes and update page backgrounds
