+++ src/client/main.tsx
@@ -6,5 +6,5 @@ import "./globals.css";
 createRoot(document.getElementById("root")!).render(
     <StrictMode>
         <App />
-    </StrictMode>
+    </StrictMode>,
 );
