+++ src/server/api/index.ts
@@ -1,0 +1,7 @@
+import { Elysia } from "elysia";
+import { authRoute } from "./auth";
+import { callbackRoute } from "./callback";
+
+export const apiRoutes = new Elysia({ prefix: "/api" })
+    .use(authRoute)
+    .use(callbackRoute);
